refactor(lib): remove top-level proxy re-export modules
