+++ src/app/stats/page.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useEffect, useMemo } from 'react';
 import { useTranslations } from '@/context/LocaleContext';
-import { usePlayerStats } from '@/lib/leveling';
+import { usePlayerStats } from '@/lib/habits/leveling';
 import { getDailyActivityHistory } from '@/lib/analytics/analytics-utils';
 import { getMissionRepository } from '@/core/repositories/mission.repository';
 import {
