+++ src/components/MissionsWidget.tsx
@@ -27,14 +27,14 @@ import { useState, useEffect } from "react";
 import { AnimatePresence } from "framer-motion";
 import { ChevronRight } from "lucide-react";
 import { Mission } from "@/data/missions";
-import { addHasanah } from "@/lib/leveling";
-import { updateStreak } from "@/lib/streak-utils";
+import { addHasanah } from "@/lib/habits/leveling";
+import { updateStreak } from "@/lib/habits/streak-utils";
 import { cn } from "@/lib/utils";
 import { usePrayerTimesContext } from "@/context/PrayerTimesContext";
 import { useMissions } from "@/hooks/useMissions";
 import MissionDetailDialog from "./MissionDetailDialog";
 import MissionListModal from "./MissionListModal";
-import { getRulingLabel } from "@/lib/mission-utils";
+import { getRulingLabel } from "@/lib/habits/mission-utils";
 import MissionSkeleton from "@/components/skeleton/MissionSkeleton";
 import { useLocale } from "@/context/LocaleContext";
 import { toast } from "sonner";
