+++ src/components/PrayerCheckInWidget.tsx
@@ -23,8 +23,8 @@ import { Check, Sparkles, AlertCircle, Calendar } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { useMissions } from "@/hooks/useMissions";
 import { usePrayerTimesContext } from "@/context/PrayerTimesContext";
-import { addHasanah } from "@/lib/leveling";
-import { updateStreak } from "@/lib/streak-utils";
+import { addHasanah } from "@/lib/habits/leveling";
+import { updateStreak } from "@/lib/habits/streak-utils";
 import { getStorageService } from "@/core/infrastructure/storage";
 import { STORAGE_KEYS } from "@/lib/constants/storage-keys";
 import { toast } from "sonner";
