--- src/lib/bookmark-storage.ts
@@ -1,1 +1,0 @@
-export * from './quran/bookmark-storage';
