--- src/lib/leveling.ts
@@ -1,1 +1,0 @@
-export * from './habits/leveling';
