--- src/lib/optimize-quran.ts
@@ -1,1 +1,0 @@
-export * from './quran/optimize-quran';
