--- src/lib/quran-utils.ts
@@ -1,1 +1,0 @@
-export * from './quran/quran-utils';
