--- src/lib/server-analytics.ts
@@ -1,1 +1,0 @@
-export * from './analytics/server-analytics';
