--- src/lib/streak-utils.ts
@@ -1,1 +1,0 @@
-export * from './habits/streak-utils';
