--- src/lib/surahData.ts
@@ -1,1 +1,0 @@
-export * from './quran/surahData';
